moved TextModal to its own file.
